requirements.txt
openai
assemblyai
streamlit
requests
python-dotenv
livekit